+++ src/main/java/com/example/encheres/dal/ArticleVenduDAO.java
@@ -11,7 +11,7 @@ public interface ArticleVenduDAO {
 	void create (ArticleVendu articleVendu);
 	ArticleVendu read (int noArticle);
 	void update (ArticleVendu articleVendu);
-	void updatePrixVente (int noArticle, float prixVente);
+	void updatePrixVente (int noArticle, int prixVente);
 	void delete (int noArticle);
 	List<ArticleVendu> findByUtilisateur(int noUtilisateur);
 	List<ArticleVendu> findAll();
